# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "swagger api",
        "title": "会议室预定 tao",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/helloworld": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试"
                ],
                "summary": "helloword",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登录详细描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"登录成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "登录失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}